requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-nmap==0.7.1
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
python-multipart==0.0.6
